# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CarModPicker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/parts/": {
      "post": {
        "tags": [
          "parts"
        ],
        "summary": "Create Part",
        "description": "Create a new shared part.",
        "operationId": "create_part_api_parts__post",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartRead"
                }
              }
            }
          },
          "400": {
            "description": "Invalid part data"
          },
          "403": {
            "description": "Not authorized to create parts"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "Read Parts",
        "description": "Get all parts with optional filtering and search.",
        "operationId": "read_parts_api_parts__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of parts to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of parts to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of parts to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of parts to return"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category ID",
              "title": "Category Id"
            },
            "description": "Filter by category ID"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in part names and descriptions",
              "title": "Search"
            },
            "description": "Search in part names and descriptions"
          }
        ],
        "responses": {
          "200": {
            "description": "List of parts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartRead"
                  },
                  "title": "Response Read Parts Api Parts  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/parts/{part_id}": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "Read Part",
        "description": "Get a specific part by ID.",
        "operationId": "read_part_api_parts__part_id__get",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Part Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartRead"
                }
              }
            }
          },
          "404": {
            "description": "Part not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "parts"
        ],
        "summary": "Update Part",
        "description": "Update a part (only creator can update).",
        "operationId": "update_part_api_parts__part_id__put",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Part Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartRead"
                }
              }
            }
          },
          "404": {
            "description": "Part not found"
          },
          "403": {
            "description": "Not authorized to update this part"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "parts"
        ],
        "summary": "Delete Part",
        "description": "Delete a part (only creator can delete).",
        "operationId": "delete_part_api_parts__part_id__delete",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Part Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartRead"
                }
              }
            }
          },
          "404": {
            "description": "Part not found"
          },
          "403": {
            "description": "Not authorized to delete this part"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/build-list-parts/build-lists/{build_list_id}/parts/{part_id}": {
      "post": {
        "tags": [
          "build_list_parts"
        ],
        "summary": "Add Part To Build List",
        "description": "Add an existing part to a build list.",
        "operationId": "add_part_to_build_list_api_build_list_parts_build_lists__build_list_id__parts__part_id__post",
        "parameters": [
          {
            "name": "build_list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build List Id"
            }
          },
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Part Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildListPartCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildListPartRead"
                }
              }
            }
          },
          "404": {
            "description": "Build list or part not found"
          },
          "403": {
            "description": "Not authorized to add parts to this build list"
          },
          "409": {
            "description": "Part already exists in build list"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "build_list_parts"
        ],
        "summary": "Update Part In Build List",
        "description": "Update a part's notes in a build list.",
        "operationId": "update_part_in_build_list_api_build_list_parts_build_lists__build_list_id__parts__part_id__put",
        "parameters": [
          {
            "name": "build_list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build List Id"
            }
          },
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Part Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildListPartUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildListPartRead"
                }
              }
            }
          },
          "404": {
            "description": "Build list or part relationship not found"
          },
          "403": {
            "description": "Not authorized to modify this build list"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "build_list_parts"
        ],
        "summary": "Remove Part From Build List",
        "description": "Remove a part from a build list.",
        "operationId": "remove_part_from_build_list_api_build_list_parts_build_lists__build_list_id__parts__part_id__delete",
        "parameters": [
          {
            "name": "build_list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build List Id"
            }
          },
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Part Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildListPartRead"
                }
              }
            }
          },
          "404": {
            "description": "Build list or part relationship not found"
          },
          "403": {
            "description": "Not authorized to modify this build list"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/build-list-parts/build-lists/{build_list_id}/parts": {
      "get": {
        "tags": [
          "build_list_parts"
        ],
        "summary": "Get Parts In Build List",
        "description": "Get all parts in a build list.",
        "operationId": "get_parts_in_build_list_api_build_list_parts_build_lists__build_list_id__parts_get",
        "parameters": [
          {
            "name": "build_list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build List Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuildListPartRead"
                  },
                  "title": "Response Get Parts In Build List Api Build List Parts Build Lists  Build List Id  Parts Get"
                }
              }
            }
          },
          "404": {
            "description": "Build list not found"
          },
          "403": {
            "description": "Not authorized to access this build list"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildListPartCreate": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "BuildListPartCreate"
      },
      "BuildListPartRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "build_list_id": {
            "type": "integer",
            "title": "Build List Id"
          },
          "part_id": {
            "type": "integer",
            "title": "Part Id"
          },
          "added_by": {
            "type": "integer",
            "title": "Added By"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "added_at": {
            "type": "string",
            "format": "date-time",
            "title": "Added At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "build_list_id",
          "part_id",
          "added_by",
          "added_at"
        ],
        "title": "BuildListPartRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PartCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "part_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Number"
          },
          "specifications": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specifications"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category_id"
        ],
        "title": "PartCreate"
      },
...
    }
  }
}